+++ source/silver/README.md
@@ -1,0 +1,11 @@
+# Silver
+
+This python package handles streaming from the bronze layer, to the silver layer.
+
+## Silver Schema
+
+See the [silver measurements](src/silver/domain/schemas/silver_measurements.py) schema for more information, on the columns and data types used in the silver layer.
+
+## Silver Transformations
+
+TBD...
